+++ pkg/plugin/plugin.go
@@ -12,10 +12,11 @@ import (
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/grafana/grafana-plugin-sdk-go/data/sqlutil"
 	"reflect"
+	"strings"
 	"time"
 )
 
-// Make sure SampleDatasource implements required interfaces. This is important to do
+// Make sure Datasource implements required interfaces. This is important to do
 // since otherwise we will only get a not implemented error response from plugin in
 // runtime. In this example datasource instance implements backend.QueryDataHandler,
 // backend.CheckHealthHandler, backend.StreamHandler interfaces. Plugin should not
@@ -25,10 +26,10 @@ import (
 // is useful to clean up resources used by previous datasource instance when a new datasource
 // instance created upon datasource settings changed.
 var (
-	_ backend.QueryDataHandler      = (*SampleDatasource)(nil)
-	_ backend.CheckHealthHandler    = (*SampleDatasource)(nil)
-	_ backend.StreamHandler         = (*SampleDatasource)(nil)
-	_ instancemgmt.InstanceDisposer = (*SampleDatasource)(nil)
+	_ backend.QueryDataHandler      = (*Datasource)(nil)
+	_ backend.CheckHealthHandler    = (*Datasource)(nil)
+	_ instancemgmt.InstanceDisposer = (*Datasource)(nil)
+	_ backend.CallResourceHandler   = (*Datasource)(nil)
 )
 
 type DatasourceSettings struct {
@@ -62,23 +63,27 @@ func NewSampleDatasource(settings backend.DataSourceInstanceSettings) (instancem
 		}
 	}
 
-	return &SampleDatasource{
+	return &Datasource{
 		databricksConnectionsString: databricksConnectionsString,
 		databricksDB:                databricksDB,
 	}, nil
 }
 
-// SampleDatasource is an example datasource which can respond to data queries, reports
+// Datasource is an example datasource which can respond to data queries, reports
 // its health and has streaming skills.
-type SampleDatasource struct {
+type Datasource struct {
 	databricksConnectionsString string
 	databricksDB                *sql.DB
+}
+
+func (d *Datasource) CallResource(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
+	return autocompletionQueries(req, sender, d.databricksDB)
 }
 
 // Dispose here tells plugin SDK that plugin wants to clean up resources when a new instance
 // created. As soon as datasource settings change detected by SDK old datasource instance will
 // be disposed and a new one will be created using NewSampleDatasource factory function.
-func (d *SampleDatasource) Dispose() {
+func (d *Datasource) Dispose() {
 	// Clean up datasource instance resources.
 }
 
@@ -86,7 +91,7 @@ func (d *SampleDatasource) Dispose() {
 // req contains the queries []DataQuery (where each query contains RefID as a unique identifier).
 // The QueryDataResponse contains a map of RefID to the response for each query, and each response
 // contains Frames ([]*Frame).
-func (d *SampleDatasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
+func (d *Datasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
 	log.DefaultLogger.Info("QueryData called", "request", req)
 
 	// create response struct
@@ -115,7 +120,7 @@ type queryModel struct {
 	QuerySettings querySettings `json:"querySettings"`
 }
 
-func (d *SampleDatasource) query(_ context.Context, pCtx backend.PluginContext, query backend.DataQuery) backend.DataResponse {
+func (d *Datasource) query(_ context.Context, pCtx backend.PluginContext, query backend.DataQuery) backend.DataResponse {
 	response := backend.DataResponse{}
 
 	// Unmarshal the JSON into our queryModel.
@@ -130,6 +135,32 @@ func (d *SampleDatasource) query(_ context.Context, pCtx backend.PluginContext, 
 	}
 
 	queryString := replaceMacros(qm.RawSqlQuery, query)
+
+	// Check if multiple statements are present in the query
+	// If so, split them and execute them individually
+	if strings.Contains(queryString, ";") {
+		// Split the query string into multiple statements
+		queries := strings.Split(queryString, ";")
+		// Check if the last statement is empty or just whitespace and newlines
+		if strings.TrimSpace(queries[len(queries)-1]) == "" {
+			// Remove the last statement
+			queries = queries[:len(queries)-1]
+		}
+		// Check if there are stil multiple statements
+		if len(queries) > 1 {
+			// Execute all but the last statement without returning any data
+			for _, query := range queries[:len(queries)-1] {
+				_, err := d.databricksDB.Exec(query)
+				if err != nil {
+					response.Error = err
+					log.DefaultLogger.Info("Error", "err", err)
+					return response
+				}
+			}
+			// Set the query string to the last statement
+			queryString = queries[len(queries)-1]
+		}
+	}
 
 	log.DefaultLogger.Info("Query", "query", queryString)
 
@@ -192,7 +223,7 @@ func (d *SampleDatasource) query(_ context.Context, pCtx backend.PluginContext, 
 // The main use case for these health checks is the test button on the
 // datasource configuration page which allows users to verify that
 // a datasource is working as expected.
-func (d *SampleDatasource) CheckHealth(_ context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
+func (d *Datasource) CheckHealth(_ context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
 	log.DefaultLogger.Info("CheckHealth called", "request", req)
 
 	dsn := d.databricksConnectionsString
@@ -218,68 +249,5 @@ func (d *SampleDatasource) CheckHealth(_ context.Context, req *backend.CheckHeal
 	return &backend.CheckHealthResult{
 		Status:  backend.HealthStatusOk,
 		Message: "Data source is working",
-	}, nil
-}
-
-// SubscribeStream is called when a client wants to connect to a stream. This callback
-// allows sending the first message.
-func (d *SampleDatasource) SubscribeStream(_ context.Context, req *backend.SubscribeStreamRequest) (*backend.SubscribeStreamResponse, error) {
-	log.DefaultLogger.Info("SubscribeStream called", "request", req)
-
-	status := backend.SubscribeStreamStatusPermissionDenied
-	if req.Path == "stream" {
-		// Allow subscribing only on expected path.
-		status = backend.SubscribeStreamStatusOK
-	}
-	return &backend.SubscribeStreamResponse{
-		Status: status,
-	}, nil
-}
-
-// RunStream is called once for any open channel.  Results are shared with everyone
-// subscribed to the same channel.
-func (d *SampleDatasource) RunStream(ctx context.Context, req *backend.RunStreamRequest, sender *backend.StreamSender) error {
-	log.DefaultLogger.Info("RunStream called", "request", req)
-
-	// Create the same data frame as for query data.
-	frame := data.NewFrame("response")
-
-	// Add fields (matching the same schema used in QueryData).
-	frame.Fields = append(frame.Fields,
-		data.NewField("time", nil, make([]time.Time, 1)),
-		data.NewField("values", nil, make([]int64, 1)),
-	)
-
-	counter := 0
-
-	// Stream data frames periodically till stream closed by Grafana.
-	for {
-		select {
-		case <-ctx.Done():
-			log.DefaultLogger.Info("Context done, finish streaming", "path", req.Path)
-			return nil
-		case <-time.After(time.Second):
-			// Send new data periodically.
-			frame.Fields[0].Set(0, time.Now())
-			frame.Fields[1].Set(0, int64(10*(counter%2+1)))
-
-			counter++
-
-			err := sender.SendFrame(frame, data.IncludeAll)
-			if err != nil {
-				log.DefaultLogger.Error("Error sending frame", "error", err)
-				continue
-			}
-		}
-	}
-}
-
-// PublishStream is called when a client sends a message to the stream.
-func (d *SampleDatasource) PublishStream(_ context.Context, req *backend.PublishStreamRequest) (*backend.PublishStreamResponse, error) {
-	log.DefaultLogger.Info("PublishStream called", "request", req)
-
-	// Do not allow publishing at all.
-	return &backend.PublishStreamResponse{
-		Status: backend.PublishStreamStatusPermissionDenied,
 	}, nil
 }