+++ src/components/QueryEditor/QueryEditor.tsx
@@ -1,20 +1,22 @@
-import { defaults } from 'lodash';
+import {defaults} from 'lodash';
 
-import React, {FormEvent, useState } from 'react';
+import React, {FormEvent, useEffect, useState} from 'react';
 import {
+    ActionMeta,
     AutoSizeInput,
-    InlineFieldRow,
-    InlineField,
-    InlineSwitch,
     CodeEditor,
     Collapse,
+    InlineField,
+    InlineFieldRow,
+    InlineSwitch, Monaco,
     Select,
-    ActionMeta,
 } from '@grafana/ui';
 import {QueryEditorProps, SelectableValue} from '@grafana/data';
 
-import { DataSource } from './datasource';
-import { defaultQuery, MyDataSourceOptions, MyQuery } from './types';
+import { editor } from 'monaco-editor/esm/vs/editor/editor.api';
+
+import {DataSource} from '../../datasource';
+import {defaultQuery, MyDataSourceOptions, MyQuery} from '../../types';
 
 type Props = QueryEditorProps<DataSource, MyQuery, MyDataSourceOptions>;
 
@@ -26,15 +28,31 @@ export function QueryEditor(props: Props) {
         { label: 'Value', value: 2, description: 'fills with a specific value' },
     ];
 
+    const [cursorPosition, setCursorPosition] = useState({lineNumber: 0, column: 0});
+    const { datasource  } = props;
+
+    const codeAutoCompletion = datasource.autoCompletionEnabled;
     const [isAdvancedOpen, setIsAdvancedOpen] = useState(false);
 
     const query = defaults(props.query, defaultQuery);
     const { rawSqlQuery, querySettings } = query;
 
+    const [queryValue, setQueryValue] = useState(rawSqlQuery || "");
+
     const onSQLQueryChange = (value: string) => {
         const { onChange, query } = props;
         onChange({ ...query, rawSqlQuery: value });
     };
+
+    const onQueryValueChange = (value: string) => {
+        setQueryValue(value);
+    }
+
+    useEffect(() => {
+        if (codeAutoCompletion) {
+            datasource.suggestionProvider.updateSuggestions(queryValue, cursorPosition);
+        }
+    }, [queryValue, cursorPosition, datasource.suggestionProvider, codeAutoCompletion]);
 
 
     const onLongToWideSwitchChange = (event: any) => {
@@ -56,6 +74,15 @@ export function QueryEditor(props: Props) {
         onChange({ ...query, querySettings: { ...querySettings, fillMode: value.value} });
     };
 
+    const getSuggestions = () => {
+        return datasource.suggestionProvider.getSuggestions();
+    }
+    const editorDidMount = async (e: editor.IStandaloneCodeEditor, m: Monaco) => {
+        e.onDidChangeCursorPosition((e) => {
+           setCursorPosition(e.position);
+        })
+    };
+
 
     // @ts-ignore
     return (
@@ -70,6 +97,9 @@ export function QueryEditor(props: Props) {
                   onSave={onSQLQueryChange}
                   showMiniMap={false}
                   showLineNumbers={false}
+                  getSuggestions={codeAutoCompletion ? getSuggestions : undefined}
+                  onChange={onQueryValueChange}
+                  onEditorDidMount={editorDidMount}
                   />
               </div>
               <Collapse label="Advanced Options" isOpen={isAdvancedOpen} onToggle={() => setIsAdvancedOpen(!isAdvancedOpen)} >